cli/requirements.txt
pyfldigi
pycryptodome
windows-curses; sys_platform == 'win32'
